+++ app/src/app/api/v1/admin/products/[id]/route.js
@@ -1,9 +1,20 @@
 import { deleteProduct, getAllProducts, getSingleProduct, updateProduct } from "@/lib/models/admin/productController";
 import { NextResponse } from "next/server";
-
+import { verifyAdminToken } from "@/lib/utils/verifyAdminToken";
 
 export async function PATCH(req, {params}) {
     try {
+        const decoded = await verifyAdminToken()
+        const {role} = decoded;
+
+        if(role === 'demo-admin'){
+            return NextResponse.json({
+                isDemo : true,
+                success: false,
+                message: "Functionality not allowed for demo admin account",
+            })
+        }
+
         const {id} = await params
         const data = await req.formData()
 
@@ -35,6 +46,17 @@ export async function PATCH(req, {params}) {
 
 export async function DELETE(req, {params}){
     try {
+        const decoded = await verifyAdminToken()
+        const {role} = decoded;
+
+        if(role === 'demo-admin'){
+            return NextResponse.json({
+                isDemo : true,
+                success: false,
+                message: "Functionality not allowed for demo admin account",
+            })
+        }
+        
         const {id} = await params
         const result = await deleteProduct(id)
